+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # Main (unreleased)
 
-- [FEATURE] New integration: grok_exporter (@roobalimsab)
 
 BREAKING CHANGE: Configuration of Tempo Autologging changed in this release.
 Please review the [migration
@@ -9,6 +8,8 @@ guide](./docs/migration-guide.md) for details.
 - [ENHANCEMENT] Add the option to log to stdout instead of a Loki instance. (@joe-elliott)
 
 - [CHANGE] Intentionally order tracing processors. (@joe-elliott)
+
+- [FEATURE] New integration: grok_exporter (@roobalimsab)
 
 # v0.14.0 (2021-05-19)